On branch TSEOB1 Changes to be committed:   (use "git reset HEAD <file>..." to unstage)
	modified:   dailycheck/daily_check.sh
	modified:   dailycheck/fileaudit/filecheck.sh
	modified:   dailycheck/fileaudit/genbas_file_attr.sh
	modified:   faillogin.sh
	modified:   lastlogin.sh
	modified:   log2csv.sh
	modified:   pwlog.sh
	modified:   remote_fast.sh
	modified:   resort.sh
	modified:   safelog.sh

+++ dailycheck/fileaudit/genbas_file_attr.sh
@@ -8,46 +8,28 @@ HOSTN=$(echo $hostname | cut -c 1-3)
 CONFIGDIR=/home/se/safechk/cfg
 SHDIR=/home/se/safechk/safesh
 case $HOSTN in
-	DAP)
-		DIR=`head -1 $CONFIGDIR/dir.conf.dap`
-		EXIST=`sed -n '2p' $CONFIGDIR/dir.conf.dap`
-		DIRNOTIME=`sed -n '3p' $CONFIGDIR/dir.conf.dap`
-		NOCHECK=`sed -n '2p' $CONFIGDIR/dir.conf.dap;sed -n '3p' $CONFIGDIR/dir.conf.dap;sed -n '4p' $CONFIGDIR/dir.conf.dap`
-		EXCLUDE=`tail -1 /home/se/safechk/cfg/dir.conf.dap | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
-		;;
-	DAR)
-		DIR=`head -1 $CONFIGDIR/dir.conf.dar`
-		EXIST=`sed -n '2p' $CONFIGDIR/dir.conf.dar`
-		DIRNOTIME=`sed -n '3p' $CONFIGDIR/dir.conf.dar`
-		NOCHECK=`sed -n '2p' $CONFIGDIR/dir.conf.dar;sed -n '3p' $CONFIGDIR/dir.conf.dar;sed -n '4p' $CONFIGDIR/dir.conf.dar`
-		EXCLUDE=`tail -1 /home/se/safechk/cfg/dir.conf.dar | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
-		;;
-	LOG)
-		DIR=`head -1 $CONFIGDIR/dir.conf.log`
-		EXIST=`sed -n '2p' $CONFIGDIR/dir.conf.log`
-		DIRNOTIME=`sed -n '3p' $CONFIGDIR/dir.conf.log`
-		NOCHECK=`sed -n '2p' $CONFIGDIR/dir.conf.log;sed -n '3p' $CONFIGDIR/dir.conf.log;sed -n '4p' $CONFIGDIR/dir.conf.log`
-		EXCLUDE=`tail -1 /home/se/safechk/cfg/dir.conf.log | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
-		;;
-	MDS)
-		DIR=`head -1 $CONFIGDIR/dir.conf.mds`
-		EXIST=`sed -n '2p' $CONFIGDIR/dir.conf.mds`
-		DIRNOTIME=`sed -n '3p' $CONFIGDIR/dir.conf.mds`
-		NOCHECK=`sed -n '2p' $CONFIGDIR/dir.conf.mds;sed -n '3p' $CONFIGDIR/dir.conf.mds;sed -n '4p' $CONFIGDIR/dir.conf.mds`
-		EXCLUDE=`tail -1 /home/se/safechk/cfg/dir.conf.mds | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
-		;;
 	WKL)
 		DIR=`head -1 $CONFIGDIR/dir.conf.wkl`
 		EXIST=`sed -n '2p' $CONFIGDIR/dir.conf.wkl`
 		DIRNOTIME=`sed -n '3p' $CONFIGDIR/dir.conf.wkl`
 		NOCHECK=`sed -n '2p' $CONFIGDIR/dir.conf.wkl;sed -n '3p' $CONFIGDIR/dir.conf.wkl;sed -n '4p' $CONFIGDIR/dir.conf.wkl`
+		NOCHECK_E=`sed -n '2p' $CONFIGDIR/dir.conf.wkl;sed -n '4p' $CONFIGDIR/dir.conf.wkl`
 		EXCLUDE=`tail -1 /home/se/safechk/cfg/dir.conf.wkl | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
+		;;
+	MIS)
+		DIR=`head -1 $CONFIGDIR/dir.conf.mis`
+		EXIST=`sed -n '2p' $CONFIGDIR/dir.conf.mis`
+		DIRNOTIME=`sed -n '3p' $CONFIGDIR/dir.conf.mis`
+		NOCHECK=`sed -n '2p' $CONFIGDIR/dir.conf.mis;sed -n '3p' $CONFIGDIR/dir.conf.mis;sed -n '4p' $CONFIGDIR/dir.conf.mis`
+		NOCHECK_E=`sed -n '2p' $CONFIGDIR/dir.conf.mis;sed -n '4p' $CONFIGDIR/dir.conf.mis`
+		EXCLUDE=`tail -1 /home/se/safechk/cfg/dir.conf.mis | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
 		;;
 	FIX)
 		DIR=`head -1 $CONFIGDIR/dir.conf.fix`
 		EXIST=`sed -n '2p' $CONFIGDIR/dir.conf.fix`
 		DIRNOTIME=`sed -n '3p' $CONFIGDIR/dir.conf.fix`
 		NOCHECK=`sed -n '2p' $CONFIGDIR/dir.conf.fix;sed -n '3p' $CONFIGDIR/dir.conf.fix;sed -n '4p' $CONFIGDIR/dir.conf.fix`
+		NOCHECK_E=`sed -n '2p' $CONFIGDIR/dir.conf.fix;sed -n '4p' $CONFIGDIR/dir.conf.fix`
 		EXCLUDE=`tail -1 /home/se/safechk/cfg/dir.conf.fix | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
 		;;
 	OTC)
@@ -55,25 +37,21 @@ case $HOSTN in
 		EXIST=`sed -n '2p' $CONFIGDIR/dir.conf.otc`
 		DIRNOTIME=`sed -n '3p' $CONFIGDIR/dir.conf.otc`
 		NOCHECK=`sed -n '2p' $CONFIGDIR/dir.conf.otc;sed -n '3p' $CONFIGDIR/dir.conf.otc;sed -n '4p' $CONFIGDIR/dir.conf.otc`
+		NOCHECK_E=`sed -n '2p' $CONFIGDIR/dir.conf.otc;sed -n '4p' $CONFIGDIR/dir.conf.otc`
 		EXCLUDE=`tail -1 /home/se/safechk/cfg/dir.conf.otc | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
-		;;
-	X23)
-		DIR=`head -1 $CONFIGDIR/dir.conf.x230`
-		EXIST=`sed -n '2p' $CONFIGDIR/dir.conf.x230`
-		DIRNOTIME=`sed -n '3p' $CONFIGDIR/dir.conf.x230`
-		NOCHECK=`sed -n '2p' $CONFIGDIR/dir.conf.x230;sed -n '3p' $CONFIGDIR/dir.conf.x230;sed -n '4p' $CONFIGDIR/dir.conf.x230`
-		EXCLUDE=`tail -1 /home/se/safechk/cfg/dir.conf.x230 | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
 		;;
 	*)
 		DIR=`head -1 $CONFIGDIR/dir.conf`
 		EXIST=`sed -n '2p' $CONFIGDIR/dir.conf`
 		DIRNOTIME=`sed -n '3p' $CONFIGDIR/dir.conf`
 		NOCHECK=`sed -n '2p' $CONFIGDIR/dir.conf;sed -n '3p' $CONFIGDIR/dir.conf;sed -n '4p' $CONFIGDIR/dir.conf`
+		NOCHECK_E=`sed -n '2p' $CONFIGDIR/dir.conf;sed -n '4p' $CONFIGDIR/dir.conf`
 		EXCLUDE=`tail -1 /home/se/safechk/cfg/dir.conf | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
 		;;
 esac
 
 ALLEXCLUDE=`echo $NOCHECK | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
+ALLEXCLUDE_E=`echo $NOCHECK_E | sed -e 's#\/#\\\/#g' -e 's/ /\/d\" -e \"\//g' -e 's/^/sed -e "\//' -e 's/$/\/d"/'`
 
 FILEDIR=/home/se/safechk/file/fileaudit
 
@@ -90,18 +68,17 @@ creat_base(){
 for DIRNAME in $DIR #import all dir_list from prompt
 do
    #${SHDIR}/sefind $DIRNAME -ls 2>/dev/null | eval $ALLEXCLUDE | sort -k2  >> $BASEFILE #generate CURRENT status to compare with BASEFILE
-	#find $DIRNAME -exec ls -cdils {} \; >> $BASEFILE 2>/dev/null
-	find $DIRNAME 2>/dev/null |xargs ls -lisd  >> $BASEFILE 2>/dev/null
-#find $DIRNAME -ls  >> $BASEFILE 2>/dev/null
-#find $DIRNAME | xargs sum >> ${BASEFILE}.sum
+	find $DIRNAME ! -type l -ls 2>/dev/null | xargs ls -lisd  >> $BASEFILE 2>/dev/null
 done
-cat $BASEFILE | eval $ALLEXCLUDE > ${BASEFILE}.tmp #generate CURRENT status to compare with BASEFILE
-mv ${BASEFILE}.tmp ${BASEFILE}
-sort -k 11 ${BASEFILE} -o ${BASEFILE}
+
+cat $BASEFILE | eval $ALLEXCLUDE > ${BASEFILE}.tmp
+mv ${BASEFILE}.tmp $BASEFILE
+sort -k 11 $BASEFILE -o $BASEFILE
 }
 
 
 creat_existbase(){
+#set -x 
 for DIRNAME in $EXIST #import all dir_list from prompt
 do
 	#${SHDIR}/sels -ld $DIRNAME  2>/dev/null | eval $EXCLUDE |awk '{print $3,$4,$1,$9}' >> $EXISTBASE
@@ -111,54 +88,41 @@ done
 for DIRNAME in $DIRNOTIME #Recursive list dir, but don't list the file and directory time. 
 do
 	#${SHDIR}/sefind $DIRNAME -ls 2>/dev/null | eval $EXCLUDE | awk '{print $5,$6,$3,$NF}'  >> $EXISTBASE
-    find $DIRNAME -ls 2>/dev/null | eval $EXCLUDE | awk '{print $5,$6,$3,$NF}'  >> $EXISTBASE
+    find $DIRNAME ! -type l -ls 2>/dev/null | eval $EXCLUDE | eval $ALLEXCLUDE_E | awk '{print $5,$6,$3,$NF}'  >> $EXISTBASE
 done
+
+sort -k 4 $EXISTBASE -o $EXISTBASE
 }
 
 
 echo Generating the Basefile for auditing
 echo ---------------------------------------------------------------
 echo IMPORTANT NOTES:
-if [ -f $BASEFILE ]; then
-	echo Basefile \"$BASEFILE\" already existed
-	echo create at \"`ls -l $BASEFILE | awk '{print $6,$7,$8,"and owner/group is",$3"/"$4}'`\"
-	echo
-	#echo If you wants to OVERWRITE it, please press \"Y\" or \"N\" to cancle
-	#echo Don\'t worry, the previous record will be backup!
-	#while [ $BOTTOM!="Y|y|N|n|Q|q" ]; do
-		#echo
-		#read BOTTOM?"Type your choise HERE "\>" "
-		#case $BOTTOM in
-                #N|n|Q|q)
-                    #echo "Basefile generate process  CANCLE by user."
-                    #exit 1 ;;
-		#Y|y)
-					mv $BASEFILE $BACKUP_BASE
-#creat_base $DIR #call function and import the parameter
-					creat_base  #call function and import the parameter
-                    mv $EXISTBASE $BACKUP_EXISTBASE
-#creat_existbase $EXIST $DIRNOTIME
-					creat_existbase #call function and import the parameter
-# creat_existbase $DIRNOTIME
-		    echo "Overwrite basefile $BASEFILE with CURRENT status, Total `awk 'END {print NR}' $BASEFILE` lines"
-		    echo "Old basefile backup to \"$BACKUP_BASE\", Total `awk 'END {print NR}' $BACKUP_BASE` lines"
-		    echo "Overwrite basefile $EXISTBASE with CURRENT status, Total `awk 'END {print NR}' $EXISTBASE` lines"
-		    echo "Old basefile backup to \"$BACKUP_EXISTBASE\", Total `awk 'END {print NR}' $BACKUP_EXISTBASE` lines"
-		    #exit 0 ;;
-		    exit 0
-		#*)
-		    #echo You enter the wrong input \"$BOTTOM\", please enter the vaild option.
-		    #echo If you wants to OVERWRITE the exist basefile, please press \"Y\" or \"N\" to cancle
-		    #;;
-		#esac
-	#done
+if [ -f $BASEFILE ] || [ -f $EXISTBASE ] ; then
+	echo Basefile \"$BASEFILE\" or  \"$EXISTBASE\" already existed
+		
+	if [ -f $BASEFILE ]; then
+		echo create at \"`ls -l $BASEFILE | awk '{print $6,$7,$8,"and owner/group is",$3"/"$4}'`\"              
+		mv $BASEFILE $BACKUP_BASE
+		creat_base
+		echo "Overwrite basefile $BASEFILE with CURRENT status, Total `awk 'END {print NR}' $BASEFILE` lines"   
+		echo "Old basefile backup to \"$BACKUP_BASE\", Total `awk 'END {print NR}' $BACKUP_BASE` lines"         
+	fi
+
+	if [ -f $EXISTBASE ]; then
+		echo create at \"`ls -l $EXISTBASE| awk '{print $6,$7,$8,"and owner/group is",$3"/"$4}'`\"
+		mv $EXISTBASE $BACKUP_EXISTBASE
+		creat_existbase
+		echo "Overwrite basefile $EXISTBASE with CURRENT status, Total `awk 'END {print NR}' $EXISTBASE` lines"
+		echo "Old basefile backup to \"$BACKUP_EXISTBASE\", Total `awk 'END {print NR}' $BACKUP_EXISTBASE` lines"
+	fi
 	exit 0
 else
 	echo "Creating basefile $BASEFILE and writing the current status in it."
-#creat_base $DIR
-#    creat_existbase $EXIST $DIRNOTIME
+	echo "Creating basefile $EXISTBASE and writing the current status in it."
 	creat_base
-    creat_existbase
+	creat_existbase
+
 	exit 0
 fi
